+++ src/index.js
@@ -3,7 +3,8 @@ import Constants from './constants'
 import Log from './log'
 import Emitter from './emitter'
 import Tracker from './tracker'
-import TrackerState from './trackerstate'
+import VideoTracker from './videotracker'
+import VideoTrackerState from './trackerstate'
 import { version } from '../package.json'
 
 export {
@@ -11,7 +12,8 @@ export {
   Log,
   Emitter,
   Tracker,
-  TrackerState,
+  VideoTracker,
+  VideoTrackerState,
   Core,
   version
 }